OOPS_SYNTAX.MD
This file contains sap abap syntax which uses oops concept for displaying and discarding data based on table data. It has checkboxes that can be disabled according to the need of user whenever the user discards the sample record. It also demonstrates functionalities of cl_gui_alv_grid classes and methods. Along with this it shows how we can directly modify or update database tables.